+++ src/Libraries Projects/Milochau.Core.Functions.Tests/TestsHelpers/TestStartup.cs
@@ -1,0 +1,6 @@
+﻿namespace Milochau.Core.Functions.Tests.TestsHelpers
+{
+    public class TestStartup : CoreFunctionsStartup
+    {
+    }
+}