+++ src/Reference Projects/Milochau.Core.Functions.ReferenceProject/Program.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Milochau.Core.Console.Infrastructure.Hosting;
 using Milochau.Core.Infrastructure.Hosting;
@@ -8,22 +7,14 @@ namespace Milochau.Core.Functions.ReferenceProject
 {
     public static class Program
     {
-        public static void Main(string[] args)
+        public static void Main()
         {
-            CreateHostBuilder(args).Build().Run();
+            CreateHostBuilder().Build().Run();
         }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
+        public static IHostBuilder CreateHostBuilder() =>
             new HostBuilder()
                 .ConfigureCoreConfiguration()
-                .ConfigureCoreHostBuilder<Startup>()
-                .ConfigureFunctionsWorkerDefaults((hostBuilderContext, functionsWorkerApplicationBuilder) =>
-                {
-                    var serviceProvider = functionsWorkerApplicationBuilder.Services.BuildServiceProvider();
-
-                    var startup = serviceProvider.GetRequiredService<CoreFunctionsStartup>();
-
-                    startup.Configure(serviceProvider, functionsWorkerApplicationBuilder);
-                });
+                .ConfigureCoreHostBuilder<Startup>();
     }
 }